fix: FaceMeasurement for Multiple Models #533
There is a misalignment of the label and highlighted area of any model that is not loaded in the center of the grid. Therefore when more than 1 model is added or the position of the model changes we get a wrong positioned label.

The reason is the position of the mesh comes from the geometry value that it does not update its position values relatively to the rest of the world.

Therefore I have added a line of code to update the selected mesh's position based on its position coordinates outside of its geometry that reference the position correctly.

Reviewd-By: Goulielmos Floros
Refs: #533