fix: upgrade three to 0.170.0 #564
Upgrade three to 0.170 at:
packages/core
packages/front

When tried to update at 0.174.0, which is the latest an error occured for importing WebGLMultipleRenderTargets as this is missing from that updated version. Hence, 0.170.0 was selected. Conscious there may be a need for to be ugraded to the other repos as well.

Reviewed-by: Goulielmos Floros
Refs: #564